fetch by cycle in bulk logger

+++ app/Http/Controllers/BulkLoggerController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Bulk_logger;
+use App\Models\RegistrationCycle;
 use App\Models\SkOfficial;
 use App\Models\User;
 use App\Models\YouthUser;
@@ -14,6 +15,11 @@ class BulkLoggerController extends Controller
     public function bulkGetBatchContent(Request $request)
     {
         //
+        $cycleID = $this->getCycle();
+
+        if (!$cycleID) {
+            return response()->json(['error' => 'No active cycle.'], 400);
+        }
         $page = $request->input('page', 1);
         $user_id = $request->input('user_id');
         User::findOrFail($user_id);
@@ -21,7 +27,10 @@ class BulkLoggerController extends Controller
         Paginator::currentPageResolver(function () use ($page) {
             return $page;
         });
-        $results = Bulk_logger::where('user_id', $user_id)->groupBy('batchNo')->paginate(10);
+        $results = Bulk_logger::whereHas('yUser', function ($q) use ($cycleID) {
+            $q->where('registration_cycle_id', $cycleID);
+        })
+            ->where('user_id', $user_id)->groupBy('batchNo')->paginate(10);
         return response()->json([
             'data' => $results->items(),
             'pagination' => [
@@ -32,6 +41,11 @@ class BulkLoggerController extends Controller
     }
     public function bulkGetUser(Request $request)
     {
+        $cycleID = $this->getCycle();
+
+        if (!$cycleID) {
+            return response()->json(['error' => 'No active cycle.'], 400);
+        }
         $page = $request->input('page', 1);
         Paginator::currentPageResolver(function () use ($page) {
             return $page;
@@ -49,23 +63,34 @@ class BulkLoggerController extends Controller
 
     public function bulkDelete(Request $request, $batchNo)
     {
+        $cycleID = $this->getCycle();
+
+        if (!$cycleID) {
+            return response()->json(['error' => 'No active cycle.'], 400);
+        }
         $bulkData = Bulk_logger::where('batchNo', $batchNo)->firstOrFail();
         Bulk_logger::destroy($bulkData->id);
         return response()->json([
             'message' => 'success'
-        ],200);
+        ], 200);
     }
     public function bulkGet(Request $request)
     {
+        $cycleID = $this->getCycle();
+
+        if (!$cycleID) {
+            return response()->json(['error' => 'No active cycle.'], 400);
+        }
         $page = $request->input('page', 1);
         $batchNo = $request->input('batchNo');
 
         Paginator::currentPageResolver(function () use ($page) {
             return $page;
         });
-        $results = YouthUser::where('batchNo', $batchNo)->with([
-            'info',
-        ])->paginate(10);
+        $results = YouthUser::where('registration_cycle_id', $cycleID)
+            ->where('batchNo', $batchNo)->with([
+                'info',
+            ])->paginate(10);
         $res = $results->items();
         $data = [];
         $full = [];
@@ -85,5 +110,10 @@ class BulkLoggerController extends Controller
                 'last_pages' => $results->lastPage(),
             ]
         ]);
+    }
+    public function getCycle()
+    {
+        $res = RegistrationCycle::where('cycleStatus', 'active')->first();
+        return $res->id ?? 0;
     }
 }
